Revert "room API 일부 (예외처리 X)"
This reverts commit aa45f66f18054ed94ce37061656a84f511ee3a32.
