+++ src/main/java/com/TeamSk/JMC/Domain/Restaurant/Restaurant.java
@@ -1,6 +1,5 @@
 package com.TeamSk.JMC.Domain.Restaurant;
 
-import com.TeamSk.JMC.Domain.Room.Room;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,10 +15,6 @@ public class Restaurant {
     @GeneratedValue
     @Column(name = "RESTAURANT_ID")
     private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "ROOM_ID")
-    private Room room;
 
     private String name;
 }
