+++ src/main/java/com/TeamSk/JMC/Domain/Room/Room.java
@@ -1,8 +1,6 @@
 package com.TeamSk.JMC.Domain.Room;
 
 import com.TeamSk.JMC.Domain.RecentRestaurant.RecentRestaurant;
-import com.TeamSk.JMC.Domain.Restaurant.Restaurant;
-import com.TeamSk.JMC.Domain.RoomMember.RoomMember;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -23,22 +21,8 @@ public class Room {
 
     private String password;
 
-    private Long leaderId;
+    private String leaderName;
 
     @OneToMany(mappedBy = "room")
     private List<RecentRestaurant> recentRestaurants;
-
-    @OneToMany(mappedBy = "room")
-    private List<RoomMember> roomMembers;
-
-    @OneToMany(mappedBy = "room")
-    private List<Restaurant> restaurants;
-
-    @Builder
-    public Room(Long id, String name, String password, Long leaderId) {
-        this.id = id;
-        this.name = name;
-        this.password = password;
-        this.leaderId = leaderId;
-    }
 }
