+++ src/main/java/com/TeamSk/JMC/Domain/Room/RoomRepository.java
@@ -1,8 +1,6 @@
 package com.TeamSk.JMC.Domain.Room;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface RoomRepository extends JpaRepository<Room, Long> {
 }
