--- src/main/java/com/TeamSk/JMC/Domain/RoomMember/RoomMemberRepository.java
@@ -1,6 +1,0 @@
-package com.TeamSk.JMC.Domain.RoomMember;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoomMemberRepository extends JpaRepository<RoomMember, Long> {
-}
