+++ src/main/java/com/TeamSk/JMC/Domain/RoomUser/RoomUserRepository.java
@@ -1,0 +1,6 @@
+package com.TeamSk.JMC.Domain.RoomUser;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoomUserRepository extends JpaRepository<RoomUser, Long> {
+}
