+++ src/main/java/com/TeamSk/JMC/Service/Login/LoginService.java
@@ -2,8 +2,10 @@ package com.TeamSk.JMC.Service.Login;
 
 
 import com.TeamSk.JMC.Domain.Member.Member;
+import com.TeamSk.JMC.Web.Dto.UserDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import java.util.Optional;
 import com.TeamSk.JMC.Domain.Member.MemberRepository;
 
 @RequiredArgsConstructor
