--- src/main/java/com/TeamSk/JMC/Service/Restaurant/RestaurantService.java
@@ -1,8 +1,0 @@
-package com.TeamSk.JMC.Service.Restaurant;
-
-import com.TeamSk.JMC.Web.Dto.roomDto.RoomResponseDto;
-
-public class RestaurantService {
-
-
-}
