+++ src/main/java/com/TeamSk/JMC/Web/Dto/MemberDto.java
@@ -1,4 +1,4 @@
-package com.TeamSk.JMC.Web.Dto.MemberDto;
+package com.TeamSk.JMC.Web.Dto;
 
 import com.TeamSk.JMC.Domain.Member.Member;
 import lombok.Getter;
@@ -10,20 +10,17 @@ public class MemberDto {
 
     private String name;
     private String email;
-    private String profileImageURL;
 
     public MemberDto(Member entity)
     {
         this.name = entity.getName();
         this.email = entity.getEmail();
-        this.profileImageURL = entity.getProfileImageURL();
     }
 
     public Member toEntity() {
         return Member.builder()
                 .name(name)
                 .email(email)
-                .profileImageURL(profileImageURL)
                 .build();
     }
 }
