+++ src/main/java/com/TeamSk/JMC/Web/Dto/RoomResponseDto.java
@@ -1,0 +1,5 @@
+package com.TeamSk.JMC.Web.Dto;
+
+public class RoomResponseDto {
+
+}
