+++ src/main/resources/static/index.html
@@ -2,12 +2,12 @@
 <html>
 <head>
     <meta charset="UTF-8">
-    <meta http-equiv="refresh" content="0; url=https://kauth.kakao.com/oauth/authorize?client_id=450bbd68a5b2bcb5e9e238e18525798a&redirect_uri=http://localhost:8080/kakao/login&response_type=code"></meta>
+    <meta http-equiv="refresh" content="0; url=https://kauth.kakao.com/oauth/authorize?client_id=7ef2e03ad8fa35a0bb154e58e9d393d7&redirect_uri=http://localhost:8080/kakao/login&response_type=code"></meta>
     <title>Insert title here</title>
 </head>
 <body>
 <!-- 카카오 로그인 -->
-<a class="p-2" href="https://kauth.kakao.com/oauth/authorize?client_id=450bbd68a5b2bcb5e9e238e18525798a&redirect_uri=http://localhost:8080/kakao/login&response_type=code">
+<a class="p-2" href="https://kauth.kakao.com/oauth/authorize?client_id=7ef2e03ad8fa35a0bb154e58e9d393d7&redirect_uri=http://localhost:8080/kakao/login&response_type=code">
     <!-- REST_API키 및 REDIRECT_URi는 본인걸로 수정하세요 -->
     <!-- 저는 redirect_uri을 http://localhost:8080/member/kakaoLogin로 했습니다. -->
     <!-- 본인걸로 수정 시 띄어쓰기 절대 하지 마세요. 오류납니다. -->
